+++ test/webhooks/helpers.js
@@ -1,3 +1,5 @@
+import HooksPostCatcher from './hooks-post-catcher.js';
+
 const helpers = {};
 
 helpers.url = 'http://127.0.0.1';
@@ -35,6 +37,43 @@ helpers.rawMessage = {
     }
   }
 };
+helpers.rawMessageUserJoined = {
+  envelope: {
+    name: 'UserJoinedMeetingEvtMsg',
+    routing: {
+      msgType: 'BROADCAST_TO_MEETING',
+      meetingId: '8043c8452ae9830aac14c517adff3839dbd9228f-1698771157700',
+      userId: 'w_xfsb9gxtlfom'
+    },
+    timestamp: 1698771164638
+  },
+  core: {
+    header: {
+      name: 'UserJoinedMeetingEvtMsg',
+      meetingId: '8043c8452ae9830aac14c517adff3839dbd9228f-1698771157700',
+      userId: 'w_xfsb9gxtlfom'
+    },
+    body: {
+      intId: 'w_xfsb9gxtlfom',
+      extId: 'w_xfsb9gxtlfom',
+      name: 'John Doe',
+      role: 'MODERATOR',
+      guest: false,
+      authed: true,
+      guestStatus: 'ALLOW',
+      emoji: 'none',
+      reactionEmoji: 'none',
+      raiseHand: false,
+      away: false,
+      pin: false,
+      presenter: false,
+      locked: true,
+      avatar: '',
+      color: '#7b1fa2',
+      clientType: 'HTML5'
+    }
+  }
+};
 
 helpers.flushall = (rClient) => {
   rClient.flushDb()
@@ -43,5 +82,17 @@ helpers.flushall = (rClient) => {
 helpers.flushredis = (hook) => {
   if (hook?.client) hook.client.flushDb();
 }
+
+helpers.createHooksCatcher = async (url, options = {}) => {
+  const catcher = new HooksPostCatcher(url, options);
+
+  await catcher.start();
+
+  return catcher;
+};
+
+helpers.stopHooksCatcher = (catcher) => {
+  if (catcher && catcher.started) catcher.stop();
+};
 
 export default helpers;